# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ffp-agent-verify",
    "description": "Verify and risk-score automated frequent-flyer award searches. See docs/ for integration guides.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/airlines": {
      "post": {
        "tags": [
          "airlines"
        ],
        "summary": "Create Airline",
        "operationId": "create_airline_v1_airlines_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirlineCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirlineCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/airlines/{airline_id}/policy": {
      "get": {
        "tags": [
          "airlines"
        ],
        "summary": "Get Policy",
        "operationId": "get_policy_v1_airlines__airline_id__policy_get",
        "parameters": [
          {
            "name": "airline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Airline Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "airlines"
        ],
        "summary": "Put Policy",
        "description": "Replace the airline's policy. Takes effect on every replica within ~1s, no redeploy.",
        "operationId": "put_policy_v1_airlines__airline_id__policy_put",
        "parameters": [
          {
            "name": "airline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Airline Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirlinePolicy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "airlines"
        ],
        "summary": "Patch Policy",
        "description": "Merge a partial policy (e.g. `{\"max_searches_per_min_per_credential\": 20}`) into the current one.",
        "operationId": "patch_policy_v1_airlines__airline_id__policy_patch",
        "parameters": [
          {
            "name": "airline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Airline Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AirlineCreate": {
        "properties": {
          "airline_id": {
            "type": "string",
            "pattern": "^[A-Z0-9]{2,3}$",
            "title": "Airline Id",
            "examples": [
              "AC"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "string"
              }
            ],
            "title": "Issuer"
          },
          "jwks": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jwks"
          },
          "jwks_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jwks Uri"
          }
        },
        "type": "object",
        "required": [
          "airline_id",
          "name",
          "issuer"
        ],
        "title": "AirlineCreate"
      },
      "AirlineCreated": {
        "properties": {
          "airline_id": {
            "type": "string",
            "title": "Airline Id"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "Shown once. Send as X-Api-Key on /v1/search/verify."
          }
        },
        "type": "object",
        "required": [
          "airline_id",
          "api_key"
        ],
        "title": "AirlineCreated"
      },
      "AirlinePolicy": {
        "properties": {
          "authorized_agents_allowed": {
            "type": "boolean",
            "title": "Authorized Agents Allowed",
            "default": true
          },
          "unauthenticated_mode": {
            "$ref": "#/components/schemas/UnauthenticatedMode",
            "default": "score"
          },
          "max_searches_per_min_per_credential": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Searches Per Min Per Credential",
            "default": 30
          },
          "max_searches_per_min_per_tool": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Searches Per Min Per Tool",
            "default": 600
          },
          "max_searches_per_min_per_ip_unauthenticated": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Searches Per Min Per Ip Unauthenticated",
            "default": 20
          },
          "max_devices_per_ip": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Devices Per Ip",
            "default": 10
          },
          "burst_multiplier": {
            "type": "number",
            "minimum": 1.0,
            "title": "Burst Multiplier",
            "default": 1.5
          },
          "min_ip_rotation_interval_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Ip Rotation Interval S",
            "default": 60
          },
          "max_concurrent_ips_per_credential": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Concurrent Ips Per Credential",
            "default": 2
          },
          "min_grant_age_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Grant Age S",
            "default": 3600
          },
          "new_grant_rate_factor": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "New Grant Rate Factor",
            "default": 0.5
          },
          "max_tools_per_member": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Tools Per Member",
            "default": 3
          },
          "vpn_tolerance": {
            "$ref": "#/components/schemas/VpnTolerance",
            "default": "authenticated_only"
          },
          "weights_authenticated": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights Authenticated"
          },
          "weights_unauthenticated": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights Unauthenticated"
          },
          "thresholds": {
            "$ref": "#/components/schemas/Thresholds"
          },
          "fail_mode": {
            "$ref": "#/components/schemas/FailMode",
            "default": "open"
          }
        },
        "type": "object",
        "title": "AirlinePolicy"
      },
...
      "FailMode": {
        "type": "string",
        "enum": [
          "open",
          "closed"
        ],
        "title": "FailMode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Thresholds": {
        "properties": {
          "challenge": {
            "type": "number",
            "title": "Challenge",
            "default": 40
          },
          "throttle": {
            "type": "number",
            "title": "Throttle",
            "default": 60
          },
          "block": {
            "type": "number",
            "title": "Block",
            "default": 80
          },
          "revoke": {
            "type": "number",
            "title": "Revoke",
            "default": 95
          }
        },
        "type": "object",
        "title": "Thresholds"
      },
...
      "UnauthenticatedMode": {
        "type": "string",
        "enum": [
          "score",
          "challenge"
        ],
        "title": "UnauthenticatedMode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VpnTolerance": {
        "type": "string",
        "enum": [
          "allow",
          "authenticated_only",
          "block"
        ],
        "title": "VpnTolerance"
      }
    }
  },
...
}